% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickchartR.R
\name{createLink}
\alias{createLink}
\title{Create Link}
\usage{
createLink(json, base64)
}
\arguments{
\item{json}{file with format Json to save in link}

\item{base64}{boolean describing if the url should or should not be encoded in base64}
}
\description{
Create Link
}
\details{
Creates link to given Json
}
\author{
Jacek Myna, Aleksandra Łuczak, Agata Pałdyna, Tomasz Radzikowski, Jan Sawicki
}
