% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickchartR.R
\name{.onAttach}
\alias{.onAttach}
\title{on attaching the package}
\usage{
.onAttach(lib, pkgname = "quickchartR")
}
\description{
on attaching the package
}
\details{
installs required packages if they are not present
}
\seealso{
rjson, caTools
}
