% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickchartR.R
\name{getDatasets}
\alias{getDatasets}
\title{From NamedList to Datasets}
\usage{
getDatasets(types, inputData, labels, colors, detailedOptions)
}
\arguments{
\item{types}{list of type of charts}

\item{inputData}{DataFrame with columns "x","y","r"}

\item{labels}{Names of columns which be used to build charts}

\item{colors}{list of background colors which be used}
}
\description{
From NamedList to Datasets
}
\details{
Prepares datasets in the form of a list to be converted to JSON
}
\author{
Jacek Myna, Aleksandra Łuczak, Agata Pałdyna, Tomasz Radzikowski, Jan Sawicki
}
