% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickchartR.R
\name{inputDataToNamedList}
\alias{inputDataToNamedList}
\title{Data To NamedList}
\usage{
inputDataToNamedList(dataFrame)
}
\arguments{
\item{dataFrame}{DataFrame with columns "x","y","r"}
}
\description{
Data To NamedList
}
\details{
Takes a data frame with columns "x", "y", "r" and returns a list that will be converted to JSON
}
\author{
Jacek Myna, Aleksandra Łuczak, Agata Pałdyna, Tomasz Radzikowski, Jan Sawicki
}
