% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickchartR.R
\docType{data}
\name{MAIN_LINK}
\alias{MAIN_LINK}
\title{Main link to quickchart.io}
\format{An object of class \code{character} of length 1.}
\usage{
MAIN_LINK
}
\description{
Main link to quickchart.io
}
\details{
Main link to quickchart.io used to construct the final link to chart
}
\keyword{datasets}
