% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickchartR.R
\name{prepareData}
\alias{prepareData}
\title{Prepare Data}
\usage{
prepareData(categories, datasets)
}
\arguments{
\item{categories}{list of categories/labels}

\item{datasets}{list of data which be used to build chart}
}
\description{
Prepare Data
}
\details{
Returns list of labels and datasets
}
\author{
Jacek Myna, Aleksandra Łuczak, Agata Pałdyna, Tomasz Radzikowski, Jan Sawicki
}
