% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickchartR.R
\name{prepareJson}
\alias{prepareJson}
\title{Prepare Json}
\usage{
prepareJson(mainType, data, options)
}
\arguments{
\item{mainType}{string that describes first (main) chart type}

\item{data}{list of datasets}

\item{options}{list of options to create Json}
}
\description{
Prepare Json
}
\details{
Functions prepare object to create Json
}
\author{
Jacek Myna, Aleksandra Łuczak, Agata Pałdyna, Tomasz Radzikowski, Jan Sawicki
}
