% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickchartR.R
\docType{data}
\name{TYPES}
\alias{TYPES}
\title{TYPES list}
\format{An object of class \code{list} of length 7.}
\usage{
TYPES
}
\description{
TYPES list
}
\details{
Defines all handled chart types
Note that type "sparkline" is not present.
In its assumption of "simplicity" which ignores labels, axis etc. sparkline chartit
is considered a terrible data visualization practice.
Even though it is usually used e.g. inside a table to minimize the eye-candy and with
a common axis with other charts, implementing it here would cause someone who is not
familiar with advanced data visualization and good practices to make a mistake (e.g. by
using it regardless of its true purpose).
It is assumed that professional data scientists do NOT use Quickcharts for their
charts because they have their own advanced tools for it.
Therefore, to somehow control and help the newbie data visualizators, sparkline chart
is NOT included in quickchartR.
}
\author{
Jacek Myna, Aleksandra Łuczak, Agata Pałdyna, Tomasz Radzikowski, Jan Sawicki
}
\keyword{datasets}
