% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickchartR.R
\name{checkTypes}
\alias{checkTypes}
\title{Check Type Function}
\usage{
checkTypes(types)
}
\arguments{
\item{types}{list of types of charts}
}
\description{
Check Type Function
}
\details{
Function checks type of charts which we want to draw.
}
\examples{
types = c("line", "bubble")
checkTypes(types)
}
\author{
Jacek Myna, Aleksandra Łuczak, Agata Pałdyna, Tomasz Radzikowski, Jan Sawicki
}
