% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickchartR.R
\name{getElementNse}
\alias{getElementNse}
\title{Get Element NSE}
\usage{
getElementNse(input_list, element_name)
}
\arguments{
\item{input_list}{object which one of elements name is element_name}

\item{element_name}{name of element without parenthesis}
}
\description{
Get Element NSE
}
\details{
Function allows to take element from the input_list without uses parenthesis
}
\author{
Jacek Myna, Aleksandra Łuczak, Agata Pałdyna, Tomasz Radzikowski, Jan Sawicki
}
