% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickchartR.R
\name{getLabels}
\alias{getLabels}
\title{Get Labels}
\usage{
getLabels(labels)
}
\arguments{
\item{labels}{list of labels}
}
\description{
Get Labels
}
\details{
Function returns list of unique labels
}
\author{
Jacek Myna, Aleksandra Łuczak, Agata Pałdyna, Tomasz Radzikowski, Jan Sawicki
}
