% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickchartR.R
\name{quickchartR}
\alias{quickchartR}
\title{hello function}
\usage{
quickchartR(types, inputData, xData, yData, colors = NULL,
  options = NULL, additionalGraphsData = NULL, base64 = T)
}
\description{
hello function
}
